+++ Documentation/CHANGELOG.md
@@ -12,8 +12,9 @@ For more information on HARK, see [our Github organization](https://github.com/e
 
 Release Date: TBD
 
-#### Major Changes
+### Major Changes
 
+- Adds a discretize method to DBlocks and RBlocks (#1460)[https://github.com/econ-ark/HARK/pull/1460]
 - Allows structural equations in model files to be provided in string form [#1427](https://github.com/econ-ark/HARK/pull/1427)
 - Introduces `HARK.parser' module for parsing configuration files into models [#1427](https://github.com/econ-ark/HARK/pull/1427)